+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@next/bundle-analyzer": "^14.2.15",
+    "@tanstack/react-query": "^5.59.16",
     "@tiptap/extension-image": "^2.8.0",
     "@tiptap/extension-link": "^2.8.0",
     "@tiptap/pm": "^2.8.0",
