+++ src/app/(home)/_components/SubBanner.tsx
@@ -6,7 +6,7 @@ export default function SubBanner() {
   const { token } = useCheckToken();
   const router = useRouter();
   return (
-    <div className="w-full h-[240px] gradient flex justify-center">
+    <div className="top-container h-[240px] gradient">
       <div className="max-w-1200 flex justify-between items-center">
         <span className="whitespace-pre-wrap text-xl font-bold text-white">
           {"문제를 풀며 정리한 내용으로\n나만의 기술 블로그를 채워보세요!"}
