+++ src/app/(sign-in)/_constants/constants.tsx
@@ -1,5 +1,26 @@
+import { AppleIcon, GoogleIcon, KakaoIcon } from "../_components/Icons";
+import { SignInBtnProps } from "../_interfaces/interfaces";
+
 export const PRIVACY_POLICY =
   "개인정보 처리 방침\n1. 목적\n본 개인정보 처리 방침은 'codingtext' 서비스(이하 '서비스')가 사용자(이하 '이용자')의 개인정보를 어떻게 수집, 이용, 보관, 보호하는지를 설명합니다. 이용자는 본 방침을 통해 자신의 개인정보가 어떻게 처리되는지를 이해할 수 있습니다.\n2. 수집하는 개인정보의 항목\n서비스는 다음과 같은 개인정보를 수집할 수 있습니다:\n이메일 주소\n사용자 ID\n비밀번호\n서비스 이용 기록\n기기 정보\n3. 개인정보 수집 및 이용 목적\n서비스는 수집한 개인정보를 다음의 목적을 위해 이용합니다:\n이용자 식별 및 인증\n서비스 제공 및 유지 관리\n이용자 맞춤형 서비스 제공\n공지사항 전달 및 고객 지원\n서비스 개선 및 새로운 서비스 개발\n4. 개인정보 보관 기간\n이용자의 개인정보는 수집 목적이 달성될 때까지 보관됩니다. 다만, 관련 법령에 의한 보관 의무가 있는 경우 해당 기간 동안 보관합니다.\n5. 개인정보의 제3자 제공\n서비스는 이용자의 개인정보를 외부에 제공하지 않습니다. 단, 법령에 따른 요청이 있을 경우 또는 이용자의 동의가 있는 경우에는 예외로 합니다.\n6. 개인정보의 안전성 확보 조치\n서비스는 이용자의 개인정보 보호를 위해 다음과 같은 조치를 취합니다:\n개인정보 암호화\n접근 통제 및 방화벽 설정\n정기적인 보안 점검\n7. 이용자의 권리\n이용자는 언제든지 자신의 개인정보 열람, 정정, 삭제 요청을 할 수 있습니다. 요청 시 신속하게 처리하겠습니다.\n8. 개인정보 처리 방침의 변경\n본 방침은 변경될 수 있으며, 변경 시 서비스 내 공지사항 또는 이메일을 통해 안내하겠습니다.\n9. 연락처\n개인정보 처리와 관련된 문의는 아래의 연락처로 문의하시기 바랍니다:\n이메일: support@codingtext.com";
 
 export const TERMS =
   "이용 약관\n제 1 조 (목적)본 약관은 'codingtext' 서비스(이하 '서비스')의 이용 조건 및 절차, 이용자와 서비스 간의 권리와 의무를 규정함을 목적으로 합니다.\n제 2 조 (약관의 효력)\n본 약관은 서비스를 이용하는 모든 이용자에게 적용됩니다.\n서비스는 필요시 약관을 변경할 수 있으며, 변경된 약관은 서비스 내 공지 또는 이메일을 통해 안내합니다.\n제 3 조 (회원가입)\n이용자는 서비스를 이용하기 위해 회원가입을 해야 합니다.\n회원가입 시 제공하는 정보는 정확하고 최신의 정보로 입력해야 하며, 허위정보를 입력할 경우 서비스 이용에 제한을 받을 수 있습니다.\n제 4 조 (서비스 이용)\n서비스는 이용자에게 프로그래밍 및 관련 콘텐츠를 제공하며, 이용자는 이를 자유롭게 이용할 수 있습니다.\n이용자는 서비스의 콘텐츠를 무단으로 복제, 배포할 수 없습니다.\n제 5 조 (이용자의 의무)\n이용자는 서비스 이용 시 다음의 행위를 하여서는 안 됩니다:\n법령에 위반되는 행위\n타인의 개인정보를 도용하거나 유출하는 행위\n서비스의 정상적인 운영을 방해하는 행위\n이용자가 본 의무를 위반할 경우, 서비스는 이용자의 이용을 제한할 수 있습니다.\n제 6 조 (서비스의 변경 및 중단)\n서비스는 기술적 사유나 운영상 필요에 따라 서비스의 일부 또는 전체를 변경하거나 중단할 수 있습니다.\n서비스 중단 시 사전에 공지하되, 긴급한 경우 사후에 공지할 수 있습니다.\n제 7 조 (책임 제한)\n서비스는 이용자가 서비스를 이용함에 있어 발생하는 손해에 대해 책임지지 않습니다.\n다만, 서비스의 고의 또는 중과실로 인한 손해는 책임을 집니다.\n제 8 조 (개인정보 보호)서비스는 이용자의 개인정보를 안전하게 보호하며, 개인정보 처리 방침에 따라 이용자의 개인정보를 관리합니다.\n제 9 조 (준거법 및 분쟁 해결)\n본 약관은 대한민국 법률에 따라 해석됩니다.\n본 약관과 관련된 분쟁은 서비스 본사 소재지의 관할 법원에 제기합니다.";
+
+export const SOCIAL_LOGIN_BUTTON_LIST: SignInBtnProps[] = [
+  {
+    icon: <GoogleIcon />,
+    service: "Google",
+    redirectionURL: "/sign-up",
+  },
+  {
+    icon: <AppleIcon />,
+    service: "Apple",
+    redirectionURL: "/sign-up",
+  },
+  {
+    icon: <KakaoIcon />,
+    service: "카카오",
+    redirectionURL: "/sign-up",
+  },
+];
