+++ src/app/(sign-in)/_interfaces/interfaces.ts
@@ -3,3 +3,20 @@ export interface SignUpForm {
   lang: string;
   profileImg: string;
 }
+
+export interface PolicyContainerProps {
+  title: string;
+  content: string;
+}
+
+export interface SignInTitleProps {
+  title: string;
+  content?: string;
+  hasBackBtn?: boolean;
+}
+
+export interface SignInBtnProps {
+  icon: JSX.Element;
+  service: string;
+  redirectionURL: string;
+}
