+++ src/app/(sign-in)/sign-up/done/page.tsx
@@ -2,6 +2,7 @@
 
 import Link from "next/link";
 import { useEffect } from "react";
+import SignInTitle from "../../_components/SignInTitle";
 
 export default function DonePage() {
   useEffect(() => {
@@ -14,12 +15,10 @@ export default function DonePage() {
 
   return (
     <div className="sign-in-container">
-      <div className="flex flex-col gap-4 items-center">
-        <span className="sign-in-title">환영합니다</span>
-        <span className="sign-in-content">
-          {"회원가입이 완료되었어요!\n즉시 서비스를 이용할 수 있어요"}
-        </span>
-      </div>
+      <SignInTitle
+        title="환영합니다"
+        content={"회원가입이 완료되었어요!\n즉시 서비스를 이용할 수 있어요"}
+      />
       <Link href="/" className="btn-primary" scroll={false}>
         메인 페이지로
       </Link>
