+++ src/app/layout.tsx
@@ -5,6 +5,8 @@ import "./globals.css";
 // PrimeReact 초기 설정
 import { PrimeReactProvider } from "primereact/api";
 import "primereact/resources/themes/lara-light-cyan/theme.css";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import QueryProvider from "./_components/QueryProvider";
 
 const pretendard = localFont({
   src: "./fonts/PretendardVariable.woff2",
@@ -23,10 +25,14 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const queryClient = new QueryClient();
+
   return (
     <html lang="en" className={`${pretendard.variable}`}>
       <body className={`${pretendard.className}`}>
-        <PrimeReactProvider>{children}</PrimeReactProvider>
+        <PrimeReactProvider>
+          <QueryProvider>{children}</QueryProvider>
+        </PrimeReactProvider>
       </body>
     </html>
   );
